backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
dnspython==2.6.1
requests==2.31.0
httpx==0.27.0
pydantic<2.0.0
cryptography==42.0.2
slowapi==0.1.9
fpdf2==2.7.5
